LLAMA-6895 : Change kwd sensitivity to a floating point
Reason for change: Change keyword sensitivity value from detector
specific integer values and ranges to flexible and consistent floating
point values. Limit checking to be handled by the detector.
Test Procedure: Test and verify sensitivity values in configuration
files are applied correctly to applicable keyword detectors. Include
maximum 1.0 and minimum 0.0 should be accepted. With sensitivity at 0.95
it, should be very difficult to trigger FFV; with sensitivity at 0.05, it
should be very easy to trigger FFV even by ambient sounds.
Risks: Low

+++ src/voice/ctrlm_voice_obj.h
@@ -173,10 +173,10 @@ typedef struct {
    guint32 keyword_sample_qty;
    guint16 doa;
    bool    standard_search_pt_triggered;
-   uint8_t standard_search_pt;
+   double  standard_search_pt;
    bool    high_search_pt_support;
    bool    high_search_pt_triggered;
-   uint8_t high_search_pt;
+   double  high_search_pt;
    double  dynamic_gain;
    struct {
       bool     selected;
@@ -291,7 +291,7 @@ typedef struct {
    unsigned long               utterance_duration_min;
    unsigned long               ffv_leading_samples;
    bool                        force_voice_settings;
-   unsigned long               keyword_sensitivity;
+   double                      keyword_sensitivity;
    bool                        vrex_test_flag;
    std::string                 opus_encoder_params_str;
    guchar                      opus_encoder_params[CTRLM_RCU_RIB_ATTR_LEN_OPUS_ENCODING_PARAMS];
